+++ CHANGELOG.md
@@ -17,3 +17,8 @@
 
 * made the core structure of simple c
 * To do: simple c and heap(probably asap)
+
+# 07/01/25
+
+* Lowkenuily finally replaced the bitmap alloc with the buddy system
+* now my ass still have the heap and the slab alloc